Add baseline with factored leniency algorithm
This tries to take a certain history of prices into account when taking BUY or SELL decisions. It also provides a shifting leniency value to ensure that the actions are taken more loosely if the initial strategy fails to decide how to react for a long time.

It aims to optimize for really big, sudden price changes, but fallback to a more lenient approach if this does not yield results for a while.